services/labeling/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
requests==2.31.0
pymongo==4.6.1
minio==7.2.0
numpy==1.24.3
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

